resolve discussion
