+++ apps/ai-dial-admin/src/components/Runs/Details/ExecutionStatusBar.tsx
@@ -2,6 +2,9 @@
 
 import { FC } from 'react';
 
+import { IconCheck, IconX } from '@tabler/icons-react';
+import classNames from 'classnames';
+
 import Grafana from '@/public/images/icons/grafana.svg';
 import { RunsI18nKey } from '@/src/constants/i18n';
 import { useI18n } from '@/src/locales/client';
@@ -28,7 +31,7 @@ const StatusPill: FC<{ status?: ExtractionResultStatus }> = ({ status }) => {
     <span
       className={`inline-flex items-center gap-1 px-2.5 py-0.5 rounded-full text-[11px] font-semibold tracking-wide ${pillClass}`}
     >
-      {isSuccess ? '\u2713' : '\u2717'} {status ?? '—'}
+      {isSuccess ? <IconCheck size={12} /> : <IconX size={12} />} {status ?? '—'}
     </span>
   );
 };
@@ -37,7 +40,7 @@ const MetaSeparator = () => <span className="w-px h-3 bg-tertiary" />;
 
 const MetaTag: FC<{ label: string; value: string; className?: string }> = ({ label, value, className }) => (
   <span className="inline-flex items-center gap-1 text-[11px] text-secondary">
-    {label} <span className={`font-medium text-primary ${className ?? ''}`}>{value}</span>
+    {label} <span className={classNames('font-medium text-primary', className)}>{value}</span>
   </span>
 );
