+++ apps/ai-dial-admin/src/components/Runs/Details/RunResultDetailPanel.tsx
@@ -13,7 +13,7 @@ import { ExtractionResult } from '@/src/models/evaluation/run';
 import AdaptiveValueGrid from './AdaptiveValueGrid';
 import CodeViewer from '@/src/components/Common/CodeViewer/CodeViewer';
 import ExecutionStatusBar from './ExecutionStatusBar';
-import { FullscreenViewerProvider } from '@/src/components/Common/FullscreenViewer/FullscreenViewer';
+import { FullscreenViewerProvider } from '@/src/context/FullscreenViewerContext';
 
 interface Props {
   result: ExtractionResult;
