+++ apps/ai-dial-admin/src/components/Runs/View/models.ts
@@ -1,6 +1,12 @@
+export interface MetricEntry {
+  key: string;
+  value: number | null;
+  isError: boolean;
+}
+
 export interface MetricGroup {
   title: string;
-  metrics: Array<{ key: string; value: number | null; isError: boolean }>;
+  metrics: MetricEntry[];
   infos?: Record<string, unknown>;
   hasError: boolean;
   errorMessage?: string;
